Finalized tester. Tuned RationalNumber constructor to allow it to be compared with RealNumber.

+++ RationalNumber.java
@@ -2,7 +2,7 @@ public class RationalNumber extends RealNumber{
     private int numerator, denominator;
 
     public RationalNumber(int nume, int deno){
-        super((double) nume/deno);
+        super( ((double) nume) / deno);
         numerator = nume;
         denominator = deno;
         if (deno == 0) {
@@ -43,9 +43,9 @@ public class RationalNumber extends RealNumber{
     }
 
     public double getValue(){
-        return numerator / denominator;
+        return ((double)numerator) / denominator;
     }
-    
+
     public int getNumerator(){
         return numerator;
     }
